gestor_ws/requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Pydantic
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP Client
httpx==0.26.0

# LangChain Core - versiones actualizadas para tool calling
langchain>=0.2.0
langchain-core>=0.2.0

# LLM - OpenAI (con soporte tool calling)
langchain-openai>=0.1.0
openai>=1.10.0

# LLM - Google Gemini (con soporte tool calling)
langchain-google-genai>=1.0.0
google-generativeai>=0.5.0

# LangGraph for Agents
langgraph>=0.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Utilities
python-dotenv==1.0.0
